refactored storage access module
