+++ auxiliary_modules/storage_access/google_cloud_storage.py
