+++ docker-compose.yml
@@ -16,7 +16,7 @@ services:
       - ./airflow_logic/:/home/airflow/gcs/airflow_logic/
       - ./map_action_logic/:/home/airflow/gcs/map_action_logic/
       - ./auxiliary_modules/api_access/:/home/airflow/gcs/api_access/
-      - ./auxiliary_modules/gcp_access/:/home/airflow/gcs/gcp_access/
+      - ./auxiliary_modules/gcp_access/:/home/airflow/gcs/gcp_settings/
       - ./auxiliary_modules/config_access/:/home/airflow/gcs/config_access/
       - ./auxiliary_modules/storage_access/:/home/airflow/gcs/storage_access/
       - ./dags_utils/:/home/airflow/gcs/dags_utils/
