+++ map_action_logic/extract/hdx_load.py
@@ -3,7 +3,7 @@ import logging
 from hdx.data.dataset import Dataset
 from hdx.hdx_configuration import Configuration
 
-from pipeline_plugin.utils.files import copy_file, save_file
+from storage_access.files import copy_file, save_file
 
 
 def get_dataset_from_hdx(hdx_address: str, dataset_name: str, output_filename: str):
