+++ map_action_logic/transform/admn_transform.py
@@ -8,8 +8,8 @@ import geopandas as gpd
 from jsonschema import validate
 
 
-from pipeline_plugin.utils.files import load_file, save_shapefiles
-from pipeline_plugin.utils.yaml_api import parse_yaml
+from storage_access.files import load_file, save_shapefiles
+from storage_access.yaml_api import parse_yaml
 
 GADM_FILENAME = "gadm36_{ISO3}.gpkg"
